requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-crispy-forms==1.7.2
django-heroku==0.3.1
gunicorn==19.9.0
heroku==0.1.4
idna==2.8
oauthlib==3.0.1
PyJWT==1.7.1
python-dateutil==1.5
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
static3==0.7.0
urllib3==1.24.1
whitenoise==4.1.2
django-rest==0.0.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0


